mediapipe_app/requirements.txt
mediapipe==0.10.21
opencv-python==4.8.1.78
numpy==1.24.3
pillow>=9.0.0